+++ click_extended/decorators/__init__.py
@@ -3,6 +3,7 @@
 from click_extended.decorators.apply import apply
 from click_extended.decorators.at_least import at_least
 from click_extended.decorators.at_most import at_most
+from click_extended.decorators.deprecated import deprecated
 from click_extended.decorators.exclusive import exclusive
 from click_extended.decorators.random_bool import random_bool
 from click_extended.decorators.random_choice import random_choice
@@ -32,6 +33,7 @@ __all__ = [
     "apply",
     "at_least",
     "at_most",
+    "deprecated",
     "exclusive",
     "random_bool",
     "random_choice",